Changed AbstractCenter and Center names, solved bug on probabilities, fixed double call to commonEndManagement

+++ code/src/main/java/it/uniroma2/pmcsn/parks/utils/RiderStatisticsWriter.java
@@ -5,11 +5,13 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Writer;
 import java.nio.file.Path;
+import java.nio.file.Files;
 
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVFormat.Builder;
 import org.apache.commons.csv.CSVPrinter;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang3.SystemUtils;
 
 import it.uniroma2.pmcsn.parks.engineering.Config;
 import it.uniroma2.pmcsn.parks.engineering.singleton.ClockHandler;
@@ -18,11 +20,14 @@ import it.uniroma2.pmcsn.parks.model.job.RiderGroup;
 public class RiderStatisticsWriter {
 
     public static void resetStatistics(String statsCase) {
-        Path statisticsDirectory = Path.of(Config.DATA_PATH, statsCase);
+        Path statisticsDirectory = Path.of(".", Config.DATA_PATH, statsCase);
 
         try {
-            FileUtils.deleteDirectory(statisticsDirectory.toFile());
-            new File(statisticsDirectory.toString()).mkdirs();
+            System.out.println(Files.exists(statisticsDirectory));
+            if (Files.exists(statisticsDirectory)) {
+                FileUtils.deleteDirectory(statisticsDirectory.toFile());
+            }
+            Files.createDirectories(statisticsDirectory);
         } catch (IOException e) {
             e.printStackTrace();
             throw new RuntimeException();
@@ -30,19 +35,21 @@ public class RiderStatisticsWriter {
 
     }
 
-    public static void writeStatistics(String fileName, RiderGroup riderGroup) {
+    public static void writeStatistics(String statsFolder, String fileName, RiderGroup riderGroup) {
 
+        Integer groupId = riderGroup.getGroupId();
         Integer groupSize = riderGroup.getGroupSize();
-        Integer priority = riderGroup.getPriority().ordinal();
+        String priority = riderGroup.getPriority().name();
         Double totalQueueTime = riderGroup.getGroupStats().getQueueTime();
         Double totalRidingTime = riderGroup.getGroupStats().getServiceTime();
         Integer totalRiding = riderGroup.getGroupStats().getTotalNumberOfVisits();
 
-        Path filePath = Path.of("Out", "Data", fileName);
-        String[] header = { "Group Size", "Priority", "Queue Time", "Riding Time", "Total Time", "Number of rides" };
+        Path filePath = Path.of(".", Config.DATA_PATH, statsFolder, fileName);
+        String[] header = { "GroupId", "GroupSize", "Priority", "QueueTime", "RidingTime", "TotalTime", "NumberRides" };
 
         // Writing the header
         if (!filePath.toFile().exists()) {
+            filePath.toFile();
             try (
                     Writer writer = new FileWriter(filePath.toFile(), true);
                     CSVPrinter csvPrinter = new CSVPrinter(writer,
@@ -59,7 +66,7 @@ public class RiderStatisticsWriter {
                 CSVPrinter csvPrinter = new CSVPrinter(writer,
                         Builder.create(CSVFormat.DEFAULT).build())) {
 
-            csvPrinter.printRecord(groupSize, priority, totalQueueTime, totalRidingTime,
+            csvPrinter.printRecord(groupId, groupSize, priority, totalQueueTime, totalRidingTime,
                     ClockHandler.getInstance().getClock(), totalRiding);
 
         } catch (IOException e) {
